Refactor to use build tools dir.
